fix process signature

+++ src/factories/calorimetry/TrackClusterMergeSplitter_factory.h
@@ -57,7 +57,7 @@ public:
   }
 
   ///! Primary algorithm call
-  void Process(int64_t /*run_number*/, uint64_t /*event_number*/) {
+  void Process(int32_t /*run_number*/, uint64_t /*event_number*/) {
     m_algo->process(
         {m_track_cluster_matches_input(), m_clusters_input(), m_track_projections_input()},
 #if EDM4EIC_VERSION_MAJOR >= 8 && EDM4EIC_VERSION_MINOR >= 4
